improve cpp_itfs utils.py file to handle readonly BUILD_DIR and add a log for pre-compile ops (#826)

+++ op_tests/test_mha.py
@@ -402,7 +402,7 @@ parser.add_argument(
     "-d",
     "--dtype",
     type=str,
-    default="bf16",
+    default="fp16",
     help="""Data type.
     e.g.: -d bf16""",
 )
